+++ proj/src/main/java/com/essentls/servlet/EventDetailServlet.java
@@ -32,9 +32,9 @@ public class EventDetailServlet extends AbstractDatabaseServlet {
         try {
             Event e = new EventInfoDAO(getConnection(),eventId).access().getOutputParam();
             HttpSession session = request.getSession();
-            long userId = -1;
+            int userId = -1;
             if(session.getAttribute("sessionUserId") != null)
-                userId = (long)session.getAttribute("sessionUserId");
+                userId = (int)session.getAttribute("sessionUserId");
             User user = new UserProfileInfoDAO(getConnection(), userId).access().getOutputParam();
             if(user == null || user.getTier() < e.getVisibility()){ //Auth check
                 request.getRequestDispatcher("/jsp/unauthorized.jsp").forward(request, response);
